refactor(mapper): merge mappers
