+++ test/unit/specs/components/report-components/ReportDescription.spec.js
