+++ test/unit/specs/components/report-components/ReportDetailsList.spec.js
