+++ test/unit/specs/components/report-components/ReportDetailsTable.spec.js
