+++ test/unit/specs/components/report-components/ReportTitle.spec.js
